+++ src/lib/pages/proposal-details/data.ts
@@ -1,6 +1,6 @@
 import { useMobile } from "lib/app-provider";
 import { useAssetInfos } from "lib/services/assetService";
-import { useMovePoolInfos } from "lib/services/move";
+import { useMovePoolInfos } from "lib/services/move/poolService";
 import { useProposalData, useProposalParams } from "lib/services/proposal";
 import type { Nullable, Option, ProposalData, ProposalParams } from "lib/types";
 import { coinToTokenWithValue, compareTokenWithValues } from "lib/utils";