+++ src/lib/pages/publish-module/utils.ts
@@ -1,4 +1,4 @@
-import type { DecodeModuleQueryResponse } from "lib/services/move/moduleService";
+import type { DecodeModuleQueryResponse } from "lib/services/types";
 import type { BechAddr20, Option } from "lib/types";
 import { UpgradePolicy } from "lib/types";
 import { bech32AddressToHex, truncate, unpadHexAddress } from "lib/utils";