+++ src/lib/pages/tx-details/components/tx-message/msg-receipts/CoinsComponent.tsx
@@ -8,7 +8,7 @@ import { ShowMoreButton } from "lib/components/button";
 import { UnsupportedTokensModal } from "lib/components/modal/UnsupportedTokensModal";
 import { TokenCard } from "lib/components/token/TokenCard";
 import { useAssetInfos } from "lib/services/assetService";
-import { useMovePoolInfos } from "lib/services/move";
+import { useMovePoolInfos } from "lib/services/move/poolService";
 import { coinToTokenWithValue, filterSupportedTokens } from "lib/utils";
 
 interface CoinsComponentProps {
@@ -37,7 +37,7 @@ export const CoinsComponent = ({ coins }: CoinsComponentProps) => {
   return (
     <Flex direction="column" w="full">
       {hasSupportedTokens && (
-        <Grid gridGap={4} gridTemplateColumns={isMobile ? "1fr " : "1fr 1fr"}>
+        <Grid gridGap={4} gridTemplateColumns={isMobile ? "1fr" : "1fr 1fr"}>
           {supportedTokens.slice(0, showMore ? undefined : 2).map((token) => (
             <TokenCard
               key={token.denom}