+++ src/lib/services/types/tx.ts
@@ -1,16 +1,24 @@
 import type { Log } from "@cosmjs/stargate/build/logs";
+import { SignMode } from "cosmjs-types/cosmos/tx/signing/v1beta1/signing";
+import type {
+  ModeInfo,
+  ModeInfo_Multi as ModeInfoMulti,
+  ModeInfo_Single as ModeInfoSingle,
+} from "cosmjs-types/cosmos/tx/v1beta1/tx";
 import { z } from "zod";
 
-import type { Transaction } from "lib/types";
+import type { BechAddr, Message, Transaction } from "lib/types";
 import {
+  ActionMsgType,
   MsgFurtherAction,
   zBechAddr,
   zCoin,
   zMessageResponse,
-  zPagination,
+  zUint8Schema,
   zUtcDate,
 } from "lib/types";
 import {
+  extractTxLogs,
   getActionMsgType,
   getMsgFurtherAction,
   parseTxHash,
@@ -22,20 +30,22 @@ import { zAny } from "./protobuf";
 // ----------------------------------------
 // --------------AuthInfo------------------
 // ----------------------------------------
-const zModeInfoMulti = z.object({
+
+let zModeInfo: z.ZodType<ModeInfo>;
+
+const zModeInfoSingle: z.ZodType<ModeInfoSingle> = z.object({
+  mode: z.custom<SignMode>((val) => SignMode[val as keyof typeof SignMode]),
+});
+
+const zModeInfoMulti: z.ZodType<ModeInfoMulti> = z.object({
   bitarray: z.object({
     extraBitsStored: z.number(),
-    elems: z.array(z.number()),
+    elems: zUint8Schema,
   }),
-  // TODO: circular dependency
-  modeInfos: z.array(z.object({})), // zModeInfo
+  modeInfos: z.lazy(() => z.array(zModeInfo)),
 });
 
-const zModeInfoSingle = z.object({
-  mode: z.string(),
-});
-
-const zModeInfo = z.object({
+zModeInfo = z.object({
   single: zModeInfoSingle.optional(),
   multi: zModeInfoMulti.optional(),
 });
@@ -121,7 +131,7 @@ const zTxResponse = z
     gas_wanted: z.string(),
     gas_used: z.string(),
     tx: zTx,
-    timestamp: zUtcDate.optional(),
+    timestamp: zUtcDate,
     events: z.array(zEvent),
   })
   .transform((val) => ({
@@ -134,6 +144,67 @@ export interface TxData extends TxResponse {
   chainId: string;
   isTxFailed: boolean;
 }
+
+export const zTxsResponseItemFromLcd = zTxResponse.transform<Transaction>(
+  (val) => {
+    const txBody = val.tx.body;
+    const message = txBody.messages[0];
+    const sender = (message?.sender ||
+      message?.signer ||
+      message?.fromAddress ||
+      "") as BechAddr;
+
+    const logs = extractTxLogs(val);
+
+    const messages = txBody.messages.map<Message>((msg, idx) => ({
+      log: logs[idx],
+      type: msg["@type"],
+    }));
+
+    return {
+      hash: val.txhash,
+      messages,
+      sender,
+      isSigner: true,
+      height: Number(val.height),
+      created: val.timestamp,
+      success: val.code === 0,
+      // TODO: implement below later
+      actionMsgType: ActionMsgType.OTHER_ACTION_MSG,
+      furtherAction: MsgFurtherAction.NONE,
+      isIbc: false,
+      isOpinit: false,
+      isInstantiate: false,
+    };
+  }
+);
+
+export const zTxsByAddressResponseLcd = z
+  .object({
+    tx_responses: z.array(zTxsResponseItemFromLcd),
+    total: z.coerce.number(),
+  })
+  .transform((val) => ({
+    items: val.tx_responses,
+    total: val.total,
+  }));
+
+export const zTxsByHashResponseLcd = z
+  .object({
+    tx_response: zTxsResponseItemFromLcd,
+  })
+  .transform((val) => ({
+    items: [val.tx_response],
+    total: 1,
+  }));
+
+export const zTxByHashResponseLcd = z
+  .object({
+    tx_response: zTxResponse,
+  })
+  .transform((val) => ({
+    txResponse: val.tx_response,
+  }));
 
 const zBaseTxsResponseItem = z.object({
   height: z.number().nonnegative(),
@@ -254,14 +325,3 @@ export const zTxsCountResponse = z
     count: z.number().nullish(),
   })
   .transform((val) => val.count);
-
-export const zTxsByAddressResponseLcd = z.object({
-  tx_responses: z.array(zTxResponse),
-  pagination: zPagination.nullable(),
-  total: z.string(),
-});
-export type TxsByAddressResponse = z.infer<typeof zTxsByAddressResponseLcd>;
-
-export const zTxByHashResponseLcd = z.object({
-  tx_response: zTxResponse,
-});